+++ snake_water_gun/menu/mode_menu.py
@@ -2,14 +2,17 @@
 import time
 import interactions as interact
 
-from snake_water_gun.lib import game_func as gm_f
 from snake_water_gun.lib import single_player as ch
+from snake_water_gun import custmizable as custm
+from snake_water_gun import basin_funct as bs_f
+
 from bot import client
-import main as m
+
+
 
 STYLE = interact.ButtonStyle.SECONDARY  # Button Style
 select_and_del = interact.api.models.message.Message()
-menu_option = ()
+
 
 
 mode_option = [
@@ -37,24 +40,34 @@ mode_option = [
 @client.component('single_player')
 async def single_player(ctx: interact.CommandContext):
     global select_and_del
-    await select_and_del.delete()
 
+    if await bs_f.user_validation(ctx, new_player=True):
+        return
+
+    await select_and_del.delete()
     ch.select_and_del = await ctx.send(components=ch.choices)
 
 
 @client.component('multi_player')
 async def multi_player(ctx: interact.CommandContext):
     global select_and_del
-    await select_and_del.delete()
 
-    msg = await ctx.send('Coming soon')
-    time.sleep(m.ERROR_MSG_INTERVEL)
+    if await bs_f.user_validation(ctx, new_player=True):
+        return
+
+    await select_and_del.delete()
+    await ctx.send('Coming soon')
+
 
 
 @client.component('back')
-async def multi_player(ctx: interact.CommandContext):
+async def back(ctx: interact.CommandContext):
     global select_and_del
+
+    if await bs_f.user_validation(ctx, new_player=True):
+        return
+
     await select_and_del.delete()  # Deleting old menu
 
-    await ctx.send('# Main Menu:-\n', components=menu_option)
-
+    from snake_water_gun.menu import main_menu as mm
+    mm.select_and_del = await ctx.send('# Menu:-\n', components=mm.menu_options)
